Add files via upload
Using the data provided in the given course, I cleaned it. For this, I checked the spelling of column L (Credit_card_type), removed empty rows and columns from the sheet, and duplicated content. However, there were also inconsistencies in the data, such as whitespace and the case of text, so I fixed them too. 

The updated file has been uploaded.