Add files via upload
In another exercise, for the month of May, I applied VLOOKUP function to find the amount of bills & utilities as shown in the file. Using HLOOKUP function, I also determined how much amount was paid in food & dining and housing from the month of January to September.
Moreover, I also applied XLOOKUP to determine the amount used for personal purpose in the month of January and April.